+++ combatsim/InfantryUnit.java
@@ -3,7 +3,7 @@
 public class InfantryUnit extends Unit {
     
     
-    public InfantryUnit(int x, int y,double z, double killProbability, double range, String team, String type){
+    public InfantryUnit(int x, int y,int z, double killProbability, double range, String team, String type){
         super(x, y,z, killProbability, range, team, type);
     }
